Add V2 ContainerInfo API
- Add V2 ContainerInfo struct
- Move v1/v2 conversion functions to a reusable package
- Add V2 ContainerInfo accessor method to Manager
