+++ api/versions_test.go
@@ -19,11 +19,9 @@ import (
 	"net/http"
 	"reflect"
 	"testing"
-	"time"
 
 	"github.com/google/cadvisor/events"
 	info "github.com/google/cadvisor/info/v1"
-	"github.com/google/cadvisor/info/v2"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -80,171 +78,4 @@ func TestGetEventRequestDoubleArgument(t *testing.T) {
 	}
 	assert.True(t, stream)
 	assert.Nil(t, err)
-}
-
-func TestInstCpuStats(t *testing.T) {
-	tests := []struct {
-		last *info.ContainerStats
-		cur  *info.ContainerStats
-		want *v2.CpuInstStats
-	}{
-		// Last is missing
-		{
-			nil,
-			&info.ContainerStats{},
-			nil,
-		},
-		// Goes back in time
-		{
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0).Add(time.Second),
-			},
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0),
-			},
-			nil,
-		},
-		// Zero time delta
-		{
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0),
-			},
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0),
-			},
-			nil,
-		},
-		// Unexpectedly small time delta
-		{
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0),
-			},
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0).Add(30 * time.Millisecond),
-			},
-			nil,
-		},
-		// Different number of cpus
-		{
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0),
-				Cpu: info.CpuStats{
-					Usage: info.CpuUsage{
-						PerCpu: []uint64{100, 200},
-					},
-				},
-			},
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0).Add(time.Second),
-				Cpu: info.CpuStats{
-					Usage: info.CpuUsage{
-						PerCpu: []uint64{100, 200, 300},
-					},
-				},
-			},
-			nil,
-		},
-		// Stat numbers decrease
-		{
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0),
-				Cpu: info.CpuStats{
-					Usage: info.CpuUsage{
-						Total:  300,
-						PerCpu: []uint64{100, 200},
-						User:   250,
-						System: 50,
-					},
-				},
-			},
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0).Add(time.Second),
-				Cpu: info.CpuStats{
-					Usage: info.CpuUsage{
-						Total:  200,
-						PerCpu: []uint64{100, 100},
-						User:   150,
-						System: 50,
-					},
-				},
-			},
-			nil,
-		},
-		// One second elapsed
-		{
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0),
-				Cpu: info.CpuStats{
-					Usage: info.CpuUsage{
-						Total:  300,
-						PerCpu: []uint64{100, 200},
-						User:   250,
-						System: 50,
-					},
-				},
-			},
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0).Add(time.Second),
-				Cpu: info.CpuStats{
-					Usage: info.CpuUsage{
-						Total:  500,
-						PerCpu: []uint64{200, 300},
-						User:   400,
-						System: 100,
-					},
-				},
-			},
-			&v2.CpuInstStats{
-				Usage: v2.CpuInstUsage{
-					Total:  200,
-					PerCpu: []uint64{100, 100},
-					User:   150,
-					System: 50,
-				},
-			},
-		},
-		// Two seconds elapsed
-		{
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0),
-				Cpu: info.CpuStats{
-					Usage: info.CpuUsage{
-						Total:  300,
-						PerCpu: []uint64{100, 200},
-						User:   250,
-						System: 50,
-					},
-				},
-			},
-			&info.ContainerStats{
-				Timestamp: time.Unix(100, 0).Add(2 * time.Second),
-				Cpu: info.CpuStats{
-					Usage: info.CpuUsage{
-						Total:  500,
-						PerCpu: []uint64{200, 300},
-						User:   400,
-						System: 100,
-					},
-				},
-			},
-			&v2.CpuInstStats{
-				Usage: v2.CpuInstUsage{
-					Total:  100,
-					PerCpu: []uint64{50, 50},
-					User:   75,
-					System: 25,
-				},
-			},
-		},
-	}
-	for _, c := range tests {
-		got, err := instCpuStats(c.last, c.cur)
-		if err != nil {
-			if c.want == nil {
-				continue
-			}
-			t.Errorf("Unexpected error: %v", err)
-		}
-		assert.Equal(t, c.want, got)
-	}
 }
